requirement.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.1
certifi
cffi
chardet==3.0.4
charset-normalizer
click==8.1.6
contourpy==1.1.1
cryptography
cycler==0.11.0
datasets==2.19.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
emoji==0.6.0
exceptiongroup==1.2.0
fastlid==0.1.9
fasttext==0.9.2
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.0
google-auth-oauthlib==1.0.0
googletrans==4.0.0rc1
gpustat==1.1
grpcio==1.58.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.23.2
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
jieba==0.42.1
joblib==1.3.2
kiwisolver==1.4.5
logzero==1.7.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mkl-fft==1.3.6
mkl-random
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.16
natsort==8.4.0
nltk==3.8.1
numpy
nvidia-ml-py==12.535.77
oauthlib==3.2.2
openai==1.12.0
openprompt==1.0.1
packaging==23.1
pandas==2.2.2
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser
pydantic==2.6.1
pydantic_core==2.16.2
pyOpenSSL
pyparsing==3.1.1
PySocks
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests
requests-oauthlib==1.3.1
rfc3986==1.5.0
rouge==1.0.0
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.96
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.66.0
transformers==4.31.0
transformers-stream-generator==0.0.4
tweet-preprocessor==0.6.0
typing_extensions
tzdata==2024.1
urllib3
wandb==0.15.12
wcwidth==0.2.6
Werkzeug==2.3.7
wordninja==2.0.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
